requirements.txt
beautifulsoup4==4.11.1
Flask==2.0.3
flask_cors==3.0.10
pymongo==3.12.0
dnspython
gunicorn==20.1.0
FPDF==1.7.2
boto3==1.23.10
mysql-connector-python==8.0.32